requirements.txt
einops
geoopt==0.5.0
h5py
loguru
matplotlib
scipy==1.13.1
numpy==1.26.4
pandas
POT
PyYAML
scikit_learn
seaborn
sinkhorn_rebasin
torch
torch_geometric
torchcfm
torchvision
tqdm
ucimlrepo
wandb
